Revamped review page permissions
