+++ src/frontend/src/components/public/PIAFormTabs/review/ProgamArea/index.tsx
@@ -6,9 +6,9 @@ import { IPiaForm } from '../../../../../types/interfaces/pia-form.interface';
 import { IReview } from '../interfaces';
 import messages from './../messages';
 import DisplayProgramArea from './displayProgramArea';
-import { statusList } from '../../../../../utils/status';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faPlus } from '@fortawesome/free-solid-svg-icons';
+import { getUserPrivileges } from '../../../../../utils/statusList/common';
 
 interface IProgramAreaProps {
   reviewForm: IReview;
@@ -28,9 +28,8 @@ const ProgramArea = (props: IProgramAreaProps) => {
     /* One can only edit Progam Area section if you are in a particular status
      * The information related to editing this section is retrieved from the statusList
      */
-    const editProgramAreaCheck = statusList(props.pia)?.[
-      Object(props.pia)?.status
-    ]?.Pages?.review?.params?.editProgramArea;
+    const editProgramAreaCheck = getUserPrivileges(props.pia)?.Pages?.review
+      ?.params?.editProgramAreaReviewers;
     setEditProgramArea(editProgramAreaCheck ?? false);
     //eslint-disable-next-line react-hooks/exhaustive-deps
   }, [props.pia?.status]);
