+++ src/frontend/src/components/public/PIAFormTabs/review/index.tsx
@@ -1,13 +1,10 @@
-import Checkbox from '../../../../components/common/Checkbox';
 import messages from './messages';
 import { ApprovalRoles, PiaStatuses } from '../../../../constant/constant';
 import { useCallback, useContext, useEffect, useMemo, useState } from 'react';
 import { IReview, IReviewSection } from './interfaces';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faPlus, faTrash } from '@fortawesome/free-solid-svg-icons';
-import { useParams } from 'react-router-dom';
-import { getGUID } from '../../../../utils/helper.util';
-import { IPiaForm } from '../../../../types/interfaces/pia-form.interface';
+import { faPlus } from '@fortawesome/free-solid-svg-icons';
+import { getGUID } from '../../../../utils/user';
 import {
   IPiaFormContext,
   PiaFormContext,
