+++ src/frontend/src/components/public/PIAFormTabs/review/viewProgramArea.tsx
@@ -4,8 +4,8 @@ import { IPiaForm } from '../../../../types/interfaces/pia-form.interface';
 import ViewReviewSection from './viewReviewSection';
 import { IReview } from './interfaces';
 import messages from './messages';
-import { getGUID } from '../../../../utils/helper.util';
-import { statusList } from '../../../../utils/status';
+import { getGUID } from '../../../../utils/user';
+import { getUserPrivileges } from '../../../../utils/statusList/common';
 
 interface IViewProgramAreaReviewProps {
   pia: IPiaForm;
@@ -46,11 +46,12 @@ const ViewProgramAreaReview = (props: IViewProgramAreaReviewProps) => {
   const handleClear = () => {
     stateChangeHandler(null, `programArea.reviews.${role}`, true);
   };
-  const canEditReviewNote =
+
+  const canEditReview =
     reviewGuid === getGUID() &&
     !printPreview &&
-    // eslint-disable-next-line @typescript-eslint/no-non-null-asserted-optional-chain
-    statusList?.(pia)?.[pia?.status!]?.Pages?.review?.params?.editReviewNote;
+    (getUserPrivileges(pia)?.Pages?.review?.params?.editProgramAreaReview ??
+      false);
 
   return (
     <div className="d-grid gap-3">
@@ -84,7 +85,7 @@ const ViewProgramAreaReview = (props: IViewProgramAreaReviewProps) => {
                   messages.PiaReviewHeader.MinistrySection.MPO.Input
                     .AcceptAccountability.en
                 }
-                canEditReview={canEditReviewNote}
+                canEditReview={canEditReview}
                 onClearClick={handleClear}
                 onConfirmClick={handleSubmit}
               />
@@ -107,7 +108,7 @@ const ViewProgramAreaReview = (props: IViewProgramAreaReviewProps) => {
               messages.PiaReviewHeader.ProgramAreaSection.Input
                 .AcceptAccountability.en
             }
-            canEditReview={canEditReviewNote}
+            canEditReview={canEditReview}
             onClearClick={handleClear}
             onConfirmClick={handleSubmit}
           />
