+++ src/frontend/src/utils/statusList/statuses/completeStatus.ts
@@ -1,0 +1,21 @@
+import { defaultCompleteModal } from '../modals';
+
+export const completeStatus = () => {
+  return {
+    title: 'Complete',
+    class: 'statusBlock__completed',
+    modal: defaultCompleteModal,
+    comments: false,
+    Pages: {
+      review: {
+        accessControl: true,
+        viewProgramAreaReviews: true,
+      },
+    },
+    Privileges: {
+      MPO: {
+        changeStatus: [],
+      },
+    },
+  };
+};
